Finished the Deck except for the drag and drop.
CardStack - we now ensure that all the cards in the stack are correctly positioned in front of each other depending on their positon in the list.

Deck - everytime the undealt cards deck is clicked, the top card is shown face up right next to the undealt deck. Once we've gone through all the cards in the undealt deck, all the cards are returned face down to the undealt deck  in the same order as before.